refactor(runtime): split host application runtime modules
