+++ apps/desktop/src-tauri/crates/host-application/src/app_service/tests/unit/mod.rs
@@ -1,0 +1,6 @@
+mod support;
+mod runtime;
+mod open_in_tools;
+mod workflow_rules;
+mod task_workflow;
+mod startup;
